+++ libsel4arm-vmm/src/arch/arm/devices/gicv2.h
@@ -1,0 +1,75 @@
+/*
+ * Copyright 2017, Data61
+ * Commonwealth Scientific and Industrial Research Organisation (CSIRO)
+ * ABN 41 687 119 230.
+ *
+ * This software may be distributed and modified according to the terms of
+ * the BSD 2-Clause license. Note that NO WARRANTY is provided.
+ * See "LICENSE_BSD2.txt" for details.
+ *
+ * @TAG(DATA61_BSD)
+ */
+
+#pragma once
+
+#include <assert.h>
+#include <stdint.h>
+
+#define GIC_DIST_PADDR       (GIC_PADDR + 0x1000)
+#define GIC_DIST_SIZE        0x1000
+#define GIC_CPU_PADDR        (GIC_PADDR + 0x2000)
+#define GIC_VCPU_CNTR_PADDR  (GIC_PADDR + 0x4000)
+#define GIC_VCPU_PADDR       (GIC_PADDR + 0x6000)
+
+/* Memory map for GICv2 distributer */
+struct gic_dist_map {
+    uint32_t enable;                /* 0x000 */
+    uint32_t ic_type;               /* 0x004 */
+    uint32_t dist_ident;            /* 0x008 */
+    uint32_t res1[29];              /* [0x00C, 0x080) */
+
+    uint32_t security[32];          /* [0x080, 0x100) */
+
+    uint32_t enable_set[32];        /* [0x100, 0x180) */
+    uint32_t enable_clr[32];        /* [0x180, 0x200) */
+    uint32_t pending_set[32];       /* [0x200, 0x280) */
+    uint32_t pending_clr[32];       /* [0x280, 0x300) */
+    uint32_t active[32];            /* [0x300, 0x380) */
+    uint32_t res2[32];              /* [0x380, 0x400) */
+
+    uint32_t priority[255];         /* [0x400, 0x7FC) */
+    uint32_t res3;                  /* 0x7FC */
+
+    uint32_t targets[255];            /* [0x800, 0xBFC) */
+    uint32_t res4;                  /* 0xBFC */
+
+    uint32_t config[64];             /* [0xC00, 0xD00) */
+
+    uint32_t spi[32];               /* [0xD00, 0xD80) */
+    uint32_t res5[20];              /* [0xD80, 0xDD0) */
+    uint32_t res6;                  /* 0xDD0 */
+    uint32_t legacy_int;            /* 0xDD4 */
+    uint32_t res7[2];               /* [0xDD8, 0xDE0) */
+    uint32_t match_d;               /* 0xDE0 */
+    uint32_t enable_d;              /* 0xDE4 */
+    uint32_t res8[70];               /* [0xDE8, 0xF00) */
+
+    uint32_t sgi_control;           /* 0xF00 */
+    uint32_t res9[3];               /* [0xF04, 0xF10) */
+    uint32_t sgi_pending_clr[4];    /* [0xF10, 0xF20) */
+    uint32_t res10[40];             /* [0xF20, 0xFC0) */
+
+    uint32_t periph_id[12];         /* [0xFC0, 0xFF0) */
+    uint32_t component_id[4];       /* [0xFF0, 0xFFF] */
+};
+
+typedef struct gic_dist_map vgic_reg_t;
+
+static inline struct gic_dist_map *priv_get_dist(void *priv)
+{
+    return (struct gic_dist_map *) priv;
+}
+
+
+extern const struct device dev_vgic_vcpu;
+extern const struct device dev_vgic_cpu;
