+++ libsel4arm-vmm/src/arch/arm/devices/vgic.c
@@ -52,99 +52,14 @@
 
 #include "../../../devices.h"
 #include "../../../vm.h"
+
 #include "virq.h"
+#include "gicv2.h"
 #include "vdist.h"
 
-#define GIC_DIST_PADDR       (GIC_PADDR + 0x1000)
-#define GIC_CPU_PADDR        (GIC_PADDR + 0x2000)
-#define GIC_VCPU_CNTR_PADDR  (GIC_PADDR + 0x4000)
-#define GIC_VCPU_PADDR       (GIC_PADDR + 0x6000)
-
-/* Memory map for GIC distributer */
-struct gic_dist_map {
-    uint32_t enable;                /* 0x000 */
-    uint32_t ic_type;               /* 0x004 */
-    uint32_t dist_ident;            /* 0x008 */
-    uint32_t res1[29];              /* [0x00C, 0x080) */
-
-    uint32_t security[32];          /* [0x080, 0x100) */
-
-    uint32_t enable_set[32];        /* [0x100, 0x180) */
-    uint32_t enable_clr[32];        /* [0x180, 0x200) */
-    uint32_t pending_set[32];       /* [0x200, 0x280) */
-    uint32_t pending_clr[32];       /* [0x280, 0x300) */
-    uint32_t active[32];            /* [0x300, 0x380) */
-    uint32_t res2[32];              /* [0x380, 0x400) */
-
-    uint32_t priority[255];         /* [0x400, 0x7FC) */
-    uint32_t res3;                  /* 0x7FC */
-
-    uint32_t targets[255];            /* [0x800, 0xBFC) */
-    uint32_t res4;                  /* 0xBFC */
-
-    uint32_t config[64];             /* [0xC00, 0xD00) */
-
-    uint32_t spi[32];               /* [0xD00, 0xD80) */
-    uint32_t res5[20];              /* [0xD80, 0xDD0) */
-    uint32_t res6;                  /* 0xDD0 */
-    uint32_t legacy_int;            /* 0xDD4 */
-    uint32_t res7[2];               /* [0xDD8, 0xDE0) */
-    uint32_t match_d;               /* 0xDE0 */
-    uint32_t enable_d;              /* 0xDE4 */
-    uint32_t res8[70];               /* [0xDE8, 0xF00) */
-
-    uint32_t sgi_control;           /* 0xF00 */
-    uint32_t res9[3];               /* [0xF04, 0xF10) */
-    uint32_t sgi_pending_clr[4];    /* [0xF10, 0xF20) */
-    uint32_t res10[40];             /* [0xF20, 0xFC0) */
-
-    uint32_t periph_id[12];         /* [0xFC0, 0xFF0) */
-    uint32_t component_id[4];       /* [0xFF0, 0xFFF] */
-};
-
-
-static inline struct gic_dist_map *vgic_priv_get_dist(struct device *d)
-{
-    assert(d);
-    assert(d->priv);
-    return vgic_device_get_vgic(d)->dist;
-}
-
-static inline void set_pending(struct gic_dist_map *gic_dist, int irq, int v)
-{
-    if (v) {
-        gic_dist->pending_set[IRQ_IDX(irq)] |= IRQ_BIT(irq);
-        gic_dist->pending_clr[IRQ_IDX(irq)] |= IRQ_BIT(irq);
-    } else {
-        gic_dist->pending_set[IRQ_IDX(irq)] &= ~IRQ_BIT(irq);
-        gic_dist->pending_clr[IRQ_IDX(irq)] &= ~IRQ_BIT(irq);
-    }
-}
-
-static inline int is_pending(struct gic_dist_map *gic_dist, int irq)
-{
-    return !!(gic_dist->pending_set[IRQ_IDX(irq)] & IRQ_BIT(irq));
-}
-
-static inline void set_enable(struct gic_dist_map *gic_dist, int irq, int v)
-{
-    if (v) {
-        gic_dist->enable_set[IRQ_IDX(irq)] |= IRQ_BIT(irq);
-        gic_dist->enable_clr[IRQ_IDX(irq)] |= IRQ_BIT(irq);
-    } else {
-        gic_dist->enable_set[IRQ_IDX(irq)] &= ~IRQ_BIT(irq);
-        gic_dist->enable_clr[IRQ_IDX(irq)] &= ~IRQ_BIT(irq);
-    }
-}
-
-static inline int is_enabled(struct gic_dist_map *gic_dist, int irq)
-{
-    return !!(gic_dist->enable_set[IRQ_IDX(irq)] & IRQ_BIT(irq));
-}
-
-static inline int is_active(struct gic_dist_map *gic_dist, int irq)
-{
-    return !!(gic_dist->active[IRQ_IDX(irq)] & IRQ_BIT(irq));
+static struct gic_dist_map *vgic_priv_get_dist(struct device *d) {
+    vgic_t *vgic = vgic_device_get_vgic(d);
+    return priv_get_dist(vgic->dist);
 }
 
 int vgic_vcpu_inject_irq(vgic_t *vgic, seL4_CPtr vcpu, struct virq_handle *irq)
@@ -185,7 +100,7 @@ int handle_vgic_maintenance(vm_t *vm, int idx)
 
     /* Clear pending */
     DIRQ("Maintenance IRQ %d\n", lr[idx]->virq);
-    set_pending(gic_dist, lr[idx]->virq, false);
+    vgic_dist_set_pending(gic_dist, lr[idx]->virq, false);
     virq_ack(lr[idx]);
 
     lr[idx] = NULL;
@@ -198,227 +113,6 @@ int handle_vgic_maintenance(vm_t *vm, int idx)
     return 0;
 }
 
-
-
-static int vgic_dist_enable(struct device *d, vm_t *vm)
-{
-    struct gic_dist_map *gic_dist = vgic_priv_get_dist(d);
-    DDIST("enabling gic distributer\n");
-    gic_dist->enable = 1;
-    return 0;
-}
-
-static int vgic_dist_disable(struct device *d, vm_t *vm)
-{
-    struct gic_dist_map *gic_dist = vgic_priv_get_dist(d);
-    DDIST("disabling gic distributer\n");
-    gic_dist->enable = 0;
-    return 0;
-}
-
-static int vgic_dist_enable_irq(struct device *d, vm_t *vm, int irq)
-{
-    struct gic_dist_map *gic_dist;
-    struct virq_handle *virq_data;
-    vgic_t *vgic;
-    gic_dist = vgic_priv_get_dist(d);
-    vgic = vgic_device_get_vgic(d);
-    DDIST("enabling irq %d\n", irq);
-    set_enable(gic_dist, irq, true);
-    virq_data = virq_find_irq_data(vgic, irq);
-    if (virq_data) {
-        /* STATE b) */
-        if (not_pending(gic_dist, virq_data->virq)) {
-            virq_ack(virq_data);
-        }
-    } else {
-        DDIST("enabled irq %d has no handle", irq);
-    }
-    return 0;
-}
-
-static int vgic_dist_disable_irq(struct device *d, vm_t *vm, int irq)
-{
-    /* STATE g) */
-    struct gic_dist_map *gic_dist = vgic_priv_get_dist(d);
-    if (irq >= 16) {
-        DDIST("disabling irq %d\n", irq);
-        set_enable(gic_dist, irq, false);
-    }
-    return 0;
-}
-
-static int vgic_dist_set_pending_irq(struct device *d, vm_t *vm, int irq)
-{
-#ifdef CONFIG_LIB_SEL4_ARM_VMM_VCHAN_SUPPORT
-    vm->lock();
-#endif //CONCONFIG_LIB_SEL4_ARM_VMM_VCHAN_SUPPORT
-
-    /* STATE c) */
-    struct gic_dist_map *gic_dist;
-    vgic_t *vgic;
-    struct virq_handle *virq_data;
-
-    gic_dist = vgic_priv_get_dist(d);
-    vgic = vgic_device_get_vgic(d);
-
-    virq_data = virq_find_irq_data(vgic, irq);
-    /* If it is enables, inject the IRQ */
-    if (virq_data && gic_dist->enable && is_enabled(gic_dist, irq)) {
-        int err;
-        DDIST("Pending set: Inject IRQ from pending set (%d)\n", irq);
-
-        set_pending(gic_dist, virq_data->virq, true);
-        err = vgic_vcpu_inject_irq(vgic, vm->vcpu.cptr, virq_data);
-        assert(!err);
-
-#ifdef CONFIG_LIB_SEL4_ARM_VMM_VCHAN_SUPPORT
-        vm->unlock();
-#endif //CONCONFIG_LIB_SEL4_ARM_VMM_VCHAN_SUPPORT
-        return err;
-    } else {
-        /* No further action */
-        DDIST("IRQ not enabled (%d)\n", irq);
-    }
-
-#ifdef CONFIG_LIB_SEL4_ARM_VMM_VCHAN_SUPPORT
-    vm->unlock();
-#endif //CONCONFIG_LIB_SEL4_ARM_VMM_VCHAN_SUPPORT
-
-    return 0;
-}
-
-static int vgic_dist_clr_pending_irq(struct device *d, vm_t *vm, int irq)
-{
-    struct gic_dist_map *gic_dist = vgic_priv_get_dist(d);
-    DDIST("clr pending irq %d\n", irq);
-    set_pending(gic_dist, irq, false);
-    return 0;
-}
-
-static int handle_vgic_dist_fault(struct device *d, vm_t *vm, fault_t *fault)
-{
-    struct gic_dist_map *gic_dist;
-    int offset;
-    enum gic_dist_action act;
-    uint32_t mask;
-    uint32_t *reg;
-
-    gic_dist = vgic_priv_get_dist(d);
-    mask = fault_get_data_mask(fault);
-    offset = fault_get_address(fault) - d->pstart;
-
-    offset = ALIGN_DOWN(offset, sizeof(uint32_t));
-
-    reg = (uint32_t *)((uintptr_t)gic_dist + offset);
-    act = gic_dist_get_action(offset);
-
-    assert(offset >= 0 && offset < d->size);
-
-    /* Out of range */
-    if (offset < 0 || offset >= sizeof(struct gic_dist_map)) {
-        DDIST("offset out of range %x %x\n", offset, sizeof(struct gic_dist_map));
-        return ignore_fault(fault);
-
-        /* Read fault */
-    } else if (fault_is_read(fault)) {
-        fault_set_data(fault, *reg);
-        return ignore_fault(fault);
-    } else {
-        uint32_t data;
-        switch (act) {
-        case ACTION_READONLY:
-            return ignore_fault(fault);
-
-        case ACTION_PASSTHROUGH:
-            *reg = fault_emulate(fault, *reg);
-            return advance_fault(fault);
-
-        case ACTION_ENABLE:
-            *reg = fault_emulate(fault, *reg);
-            data = fault_get_data(fault);
-            if (data == 1) {
-                vgic_dist_enable(d, vm);
-            } else if (data == 0) {
-                vgic_dist_disable(d, vm);
-            } else {
-                assert(!"Unknown enable register encoding\n");
-            }
-            return advance_fault(fault);
-
-        case ACTION_ENABLE_SET:
-            data = fault_get_data(fault);
-            /* Mask the data to write */
-            data &= mask;
-            /* Mask bits that are already set */
-            data &= ~(*reg);
-            while (data) {
-                int irq;
-                irq = CTZ(data);
-                data &= ~(1U << irq);
-                irq += (offset - 0x100) * 8;
-                vgic_dist_enable_irq(d, vm, irq);
-            }
-            return ignore_fault(fault);
-
-        case ACTION_ENABLE_CLR:
-            data = fault_get_data(fault);
-            /* Mask the data to write */
-            data &= mask;
-            /* Mask bits that are already clear */
-            data &= *reg;
-            while (data) {
-                int irq;
-                irq = CTZ(data);
-                data &= ~(1U << irq);
-                irq += (offset - 0x180) * 8;
-                vgic_dist_disable_irq(d, vm, irq);
-            }
-            return ignore_fault(fault);
-
-        case ACTION_PENDING_SET:
-            data = fault_get_data(fault);
-            /* Mask the data to write */
-            data &= mask;
-            /* Mask bits that are already set */
-            data &= ~(*reg);
-            while (data) {
-                int irq;
-                irq = CTZ(data);
-                data &= ~(1U << irq);
-                irq += (offset - 0x200) * 8;
-                vgic_dist_set_pending_irq(d, vm, irq);
-            }
-            return ignore_fault(fault);
-
-        case ACTION_PENDING_CLR:
-            data = fault_get_data(fault);
-            /* Mask the data to write */
-            data &= mask;
-            /* Mask bits that are already clear */
-            data &= *reg;
-            while (data) {
-                int irq;
-                irq = CTZ(data);
-                data &= ~(1U << irq);
-                irq += (offset - 0x280) * 8;
-                vgic_dist_clr_pending_irq(d, vm, irq);
-            }
-            return ignore_fault(fault);
-
-        case ACTION_SGI:
-            assert(!"vgic SGI not implemented!\n");
-            return ignore_fault(fault);
-
-        case ACTION_UNKNOWN:
-        default:
-            DDIST("Unknown action on offset 0x%x\n", offset);
-            return ignore_fault(fault);
-        }
-    }
-    abandon_fault(fault);
-    return -1;
-}
 
 static void vgic_dist_reset(struct device *d)
 {
@@ -482,7 +176,8 @@ int vm_inject_IRQ(virq_handle_t virq)
         return -1;
     }
 
-    vgic_dist_set_pending_irq(vgic_device, vm, virq->virq);
+    vgic_t *vgic = vgic_device_get_vgic(vgic_device);
+    vgic_dist_set_pending_irq(vgic, vm->vcpu.cptr, virq->virq);
 
     if (!fault_handled(vm->fault) && fault_is_wfi(vm->fault)) {
         ignore_fault(vm->fault);
