+++ libsel4arm-vmm/src/arch/arm/devices/vgic.h
@@ -11,7 +11,5 @@
  */
 #include "../../../vm.h"
 
-extern const struct device dev_vgic_vcpu;
-extern const struct device dev_vgic_cpu;
 
 int handle_vgic_maintenance(vm_t *vm, int idx);
